backend-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
pydantic[email]==2.5.0
email-validator==2.1.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt>=4.0,<5.0
python-multipart==0.0.6
redis==5.0.1
alembic==1.12.1
